Workspace cleanup: bitwrk-server
- Use go modules
- Move bitwrk-server to /server
