+++ go.mod
@@ -3,6 +3,8 @@ module github.com/indyjo/bitwrk
 go 1.12
 
 require (
+	bitbucket.org/ww/goautoneg v0.0.0-20120707110453-75cd24fc2f2c
 	github.com/indyjo/bitwrk-common v0.0.0-20190807174825-b6eccd2e81a2
 	github.com/indyjo/cafs v0.0.0-20170618142142-82b9e9a1a07f
+	google.golang.org/appengine v1.6.1
 )
