+++ server/appengine/js/createbid.js
