+++ server/appengine/js/createmessage.js
