+++ server/appengine/server.go
@@ -1,7 +1,7 @@
 package main
 
 import (
-	_ "github.com/indyjo/bitwrk-server"
+	_ "github.com/indyjo/bitwrk/server"
 	"google.golang.org/appengine"
 )
