+++ server/query/query_accounts.go
@@ -17,18 +17,20 @@
 package query
 
 import (
-	"encoding/json"
-	db "github.com/indyjo/bitwrk-server/gae"
-	"google.golang.org/appengine"
-	"google.golang.org/appengine/log"
+	"bytes"
+	"fmt"
 	"net/http"
 	"strconv"
-	"time"
+
+	db "github.com/indyjo/bitwrk/server/gae"
+	"google.golang.org/appengine"
+	"google.golang.org/appengine/log"
 )
 
-// Handles requests for account movements (ledger entries)
-func HandleQueryAccountMovements(w http.ResponseWriter, r *http.Request) {
+// Handles requests for sets of account IDs.
+func HandleQueryAccounts(w http.ResponseWriter, r *http.Request) {
 	c := appengine.NewContext(r)
+
 	limitStr := r.FormValue("limit")
 	var limit int
 	if limitStr == "" {
@@ -40,23 +42,19 @@ func HandleQueryAccountMovements(w http.ResponseWriter, r *http.Request) {
 		limit = int(n)
 	}
 
-	beginStr := r.FormValue("begin")
-	var begin time.Time
-	if beginStr != "" {
-		if t, err := time.Parse(time.RFC3339, beginStr); err != nil {
-			http.Error(w, err.Error(), http.StatusInternalServerError)
-			return
-		} else {
-			begin = t
-		}
+	requestdepositaddress := r.FormValue("requestdepositaddress") != ""
+
+	buffer := new(bytes.Buffer)
+	handler := func(key string) {
+		fmt.Fprintf(buffer, "%v\n", key)
 	}
 
-	if result, err := db.QueryAccountMovements(c, begin, limit); err != nil {
-		log.Errorf(c, "QueryAccountMovements failed: %v", err)
+	if err := db.QueryAccountKeys(c, limit, requestdepositaddress, handler); err != nil {
+		log.Errorf(c, "QueryAccountKeys failed: %v", err)
 		http.Error(w, err.Error(), http.StatusInternalServerError)
 		return
-	} else {
-		w.Header().Set("Content-Type", "application/json")
-		json.NewEncoder(w).Encode(result)
 	}
+
+	w.Header().Set("Content-Type", "text/plain")
+	buffer.WriteTo(w)
 }
