+++ server/server_motd.go
