+++ server/server_nonce.go
@@ -20,11 +20,12 @@ import (
 	"context"
 	"crypto/md5"
 	"fmt"
+	"net/http"
+	"time"
+
 	"google.golang.org/appengine"
 	"google.golang.org/appengine/datastore"
 	"google.golang.org/appengine/log"
-	"net/http"
-	"time"
 )
 
 type Nonce struct {
